Вынес в отдельный файл кнопку выбора товара
